simplify struture to combine server and client into root.
