+++ package.json
@@ -1,0 +1,31 @@
+{
+    "name": "sliding-puzzle",
+    "version": "0.1.0",
+    "type": "module",
+    "main": "index.js",
+    "private": true,
+    "scripts": {
+        "start": "node server.js",
+        "dev": "nodemon server.js",
+        "client": "vite",
+        "build": "vite build",
+        "preview": "vite preview"
+    },
+    "dependencies": {
+        "@vitejs/plugin-react": "^5.0.4",
+        "bcrypt": "^5.1.0",
+        "cors": "^2.8.5",
+        "dotenv": "^16.0.3",
+        "express": "^4.18.2",
+        "jsonwebtoken": "^9.0.0",
+        "mongodb": "^6.0.0",
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
+        "vite": "^7.1.9",
+        "vite-express": "^0.21.1"
+    },
+    "devDependencies": {
+        "@vitejs/plugin-react": "^4.0.0",
+        "nodemon": "^2.0.22"
+    }
+}
