--- server/package.json
@@ -1,21 +1,0 @@
-{
-    "name": "sliding-puzzle-server",
-    "version": "0.1.0",
-    "type": "module",
-    "main": "index.js",
-    "scripts": {
-        "start": "node index.js",
-        "dev": "nodemon index.js"
-    },
-    "dependencies": {
-        "bcrypt": "^5.1.0",
-        "cors": "^2.8.5",
-        "dotenv": "^16.0.3",
-        "express": "^4.18.2",
-        "jsonwebtoken": "^9.0.0",
-        "mongodb": "^6.0.0"
-    },
-    "devDependencies": {
-        "nodemon": "^2.0.22"
-    }
-}
