+++ src/components/SlidingPuzzle.jsx
