+++ vite.config.js
@@ -2,8 +2,5 @@ import { defineConfig } from 'vite'
 import react from '@vitejs/plugin-react'
 
 export default defineConfig({
-    plugins: [react()],
-    server: {
-        port: 3000,
-    }
+    plugins: [react()]
 })
